# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The signed-in user, from the BetterAuth session cookie."
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/hotels": {
      "get": {
        "operationId": "HotelsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive filter matched against name and location",
            "schema": {
              "maxLength": 100,
              "example": "Barcelona",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 80,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedHotelsDto"
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
    "/hotels/within-bounds": {
      "get": {
        "operationId": "HotelsController_findWithinBounds",
        "parameters": [
          {
            "name": "swLat",
            "required": true,
            "in": "query",
            "description": "South-west corner latitude",
            "schema": {
              "example": 41,
              "type": "number"
            }
          },
          {
            "name": "swLng",
            "required": true,
            "in": "query",
            "description": "South-west corner longitude",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "neLat",
            "required": true,
            "in": "query",
            "description": "North-east corner latitude",
            "schema": {
              "example": 41.6,
              "type": "number"
            }
          },
          {
            "name": "neLng",
            "required": true,
            "in": "query",
            "description": "North-east corner longitude",
            "schema": {
              "example": 2.4,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
    "/hotels/nearest": {
      "get": {
        "operationId": "HotelsController_findNearest",
        "parameters": [
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "description": "Longitude",
            "schema": {
              "example": 2.17,
              "type": "number"
            }
          },
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "description": "Latitude",
            "schema": {
              "example": 41.38,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
    "/hotels/{id}": {
      "get": {
        "operationId": "HotelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDto"
                }
              }
            }
          },
          "400": {
            "description": "id is not a valid UUID"
          },
          "404": {
            "description": "Hotel with the given id does not exist"
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
    "/bookings": {
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [
          {
            "name": "x-synthetic-booking",
            "required": false,
            "in": "header",
            "description": "Set to \"true\" to tag this booking as test data created by an automated e2e suite, rather than a real reservation. Synthetic bookings are excluded from nothing at read time — they behave like any other booking — but can be bulk-deleted via DELETE /bookings/synthetic.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        },
        "tags": [
          "bookings"
        ]
      },
      "get": {
        "operationId": "BookingsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDetailsDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/hotel/{hotelId}": {
      "get": {
        "operationId": "BookingsController_findCurrentByHotel",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "id is not a valid UUID"
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/user/{userId}": {
      "get": {
        "operationId": "BookingsController_getBookingsByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDetailsDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found"
          }
        },
        "tags": [
          "bookings"
        ]
      },
      "patch": {
        "operationId": "BookingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "bookings"
        ]
      },
      "delete": {
        "operationId": "BookingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Booking not found"
          }
        },
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/synthetic": {
      "delete": {
        "operationId": "BookingsController_removeSynthetic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSyntheticBookingsDto"
                }
              }
            }
          }
        },
        "tags": [
          "bookings"
        ]
      }
    }
  },
  "info": {
    "title": "Hotels API",
    "description": "Public API contract for the Hotels app",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "GeoDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 41.3851,
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "example": 2.1734,
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "HotelImageDto": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "hotels/shared/01.jpg",
            "description": "Storage key, relative to the configured storage/CDN base URL. The client resolves it as `${STORAGE_URL}/${path}`."
          },
          "alt": {
            "type": "string",
            "example": "Hotel Barcino Central — photo 1"
          },
          "width": {
            "type": "number",
            "example": 1600,
            "description": "Intrinsic width in pixels."
          },
          "height": {
            "type": "number",
            "example": 1067,
            "description": "Intrinsic height in pixels."
          }
        },
        "required": [
          "path",
          "alt",
          "width",
          "height"
        ]
      },
      "HotelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Hotel Barcino Central"
          },
          "description": {
            "type": "string",
            "example": "A boutique hotel steps away from Las Ramblas, blending Gothic Quarter charm with modern comfort."
          },
          "location": {
            "type": "string",
            "example": "Barcelona, Spain"
          },
          "geo": {
            "$ref": "#/components/schemas/GeoDto"
          },
          "images": {
            "description": "Gallery images for the hotel, in display order. First entry is the cover.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelImageDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "location",
          "geo",
          "images"
        ]
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "pageCount": {
            "type": "number",
            "example": 5
          },
          "total": {
            "type": "number",
            "example": 87
          }
        },
        "required": [
          "page",
          "pageSize",
          "pageCount",
          "total"
        ]
      },
      "HotelsMetaDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": [
          "pagination"
        ]
      },
      "PaginatedHotelsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotelDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/HotelsMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "example": "38dca5bd-0417-4971-baee-056e1aa3ce21"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-09-01"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-09-10"
          }
        },
        "required": [
          "hotelId",
          "checkIn",
          "checkOut"
        ]
      },
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "38dca5bd-0417-4971-baee-056e1aa3ce21"
          },
          "userId": {
            "type": "string",
            "example": "bf721a73-1a8b-4de2-b74b-a747e1197d3f"
          },
          "hotelId": {
            "type": "string",
            "example": "38dca5bd-0417-4971-baee-056e1aa3ce21"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-09-01"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-09-10"
          }
        },
        "required": [
          "id",
          "userId",
          "hotelId",
          "checkIn",
          "checkOut"
        ]
      },
      "BookingHotelSummaryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "BookingUserSummaryDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "BookingDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hotel": {
            "$ref": "#/components/schemas/BookingHotelSummaryDto"
          },
          "user": {
            "$ref": "#/components/schemas/BookingUserSummaryDto"
          },
          "checkIn": {
            "type": "string"
          },
          "checkOut": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "hotel",
          "user",
          "checkIn",
          "checkOut"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "string",
            "example": "38dca5bd-0417-4971-baee-056e1aa3ce21"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-09-01"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-09-10"
          }
        }
      },
      "DeleteSyntheticBookingsDto": {
        "type": "object",
        "properties": {
          "deletedCount": {
            "type": "number",
            "example": 3,
            "description": "Number of synthetic bookings deleted"
          }
        },
        "required": [
          "deletedCount"
        ]
      }
    }
  }
}